pyproject.toml
[project]
name = "gigabot"
version = "1.0.0"
description = "Enterprise-grade AI assistant framework with multi-channel support, tiered routing, and swarm orchestration"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GigaBot contributors"}
]
keywords = ["ai", "agent", "chatbot", "assistant", "multi-agent", "swarm", "telegram", "discord", "slack"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
browser = [
    "playwright>=1.40.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
]
discord = [
    "discord.py>=2.3.0",
]
matrix = [
    "matrix-nio>=0.24.0",
]
slack = [
    "slack-bolt>=1.18.0",
]
tiktoken = [
    "tiktoken>=0.5.0",
]
all = [
    "gigabot[browser,embeddings,discord,matrix,slack,tiktoken]",
]

[project.scripts]
nanobot = "nanobot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot"]

[tool.hatch.build.targets.wheel.sources]
"nanobot" = "nanobot"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "nanobot/**/*.py",
    "nanobot/skills/**/*.md",
    "nanobot/skills/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "nanobot/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "nanobot/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
